this is the latest version of my 2 yr + project ... bear with me
this is what i've worked up to over the years. Its very rough, and this
project has helped me learn a lot about python and programming, as i'm
sure an experienced coder can tell through the differences in
sophistication and complexity in newer changes versus older ones. I
wrote this with the mindset of a writer, that I would make it functional
first, and then edit it later, for cleanliness and optimization.